requirements.txt
yfinance>=0.2.0
pandas>=2.0.0
numpy>=1.22.0
requests>=2.28.0
curl_cffi>=0.5.0
aiogram==3.19.0
aiohttp>=3.9.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
